Omit software. Create 3D images of scrolls & text
Omit helps you to create 3D files of scrolls on which text can be read.
The omit.m file is a matlab file. You can open it in the matlab trial version and access the code.
The omit.exe file is the compiled matlab code. 
Then you do not have to install Matlab itself, but except for some settings, you cannot alter the code